+++ examples/examples_test.go
@@ -96,6 +96,21 @@ func TestNginxPy(t *testing.T) {
 	integration.ProgramTest(t, &opts)
 }
 
+func TestNginxCs(t *testing.T) {
+	cwd, err := os.Getwd()
+	if !assert.NoError(t, err) {
+		t.FailNow()
+	}
+
+	opts := base.With(integration.ProgramTestOptions{
+		Dependencies: []string{
+			"Pulumi.Docker",
+		},
+		Dir: path.Join(cwd, "nginx-cs"),
+	})
+	integration.ProgramTest(t, &opts)
+}
+
 func TestDotNet(t *testing.T) {
 	cwd, err := os.Getwd()
 	if !assert.NoError(t, err) {